# documentation/api.yaml
openapi: 3.0.0
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0
  licence:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: http://localhost:5000
    description: Dev Server on Port
  # - url: http://staging-api.example.com
  #   description: Optional server description, e.g. Internal staging server for testing
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
security:
  - bearerAuth: []
paths:
...
  /admin/lessson/create:
    post:
      summary: Admin Add Lesson create
      tags:
        - Admin Course
      description: Admin upload Lesson
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                courseId:
                  type: string
                adminId:
                  type: string
                description:
                  type: string
                lesson:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "admin created lesson sucsess fully"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/lesson/update/{id}:
    put:
      summary: Admin course Lesson update
      tags:
        - Admin Course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the lesson ID to Admin updated lesson
          example: 62b216049920204eb56c309b
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                lesson:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "admin update lesson sucsess fully"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/lesson/delete/{id}:
    delete:
      put:
      summary: Admin delete course lesson
      tags:
        - Admin Course
      description: Admin Delete course lesson
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the lesson ID to Admin deleted course lesson
          example: 62aec016a745eeaa456ff998
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                deletedCourse: string
              example:
                deletedCourse: "User sucsess fuly deleted cours comment"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/courseBooks/create:
    post:
      summary: Admin Add Lesson create
      tags:
        - Admin Course
      description: Admin upload Lesson
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                courseId:
                  type: string
                urlBooks:
                  type: string
                description:
                  type: string
                books:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "admin created books sucsess fully"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/coursBooks/update/{id}:
    put:
      summary: Admin course book update
      tags:
        - Admin Course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the book ID to Admin updated book
          example: 62b2328a357159caf671f4cd
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                urlBooks:
                  type: string
                description:
                  type: string
                books:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "admin update books sucsess fully"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/coursBooks/delete/{id}:
      delete:
        put:
        summary: Admin delete course book
        tags:
          - Admin Course
        description: Admin Delete course book
        parameters:
          - in: path
            name: id
            schema:
              type: string
            required: true
            description: this is the book ID to Admin deleted course book
            example: 62aec016a745eeaa456ff998
        responses:
          200:
            description: OK
            content:
              application/json:
                schema:
                  tyep: object
                  properties:
                  deletedCourse: string
                example:
                  deletedCourse: "Admin sucsess fuly deleted cours book"
          404:
            description: Not Found Course
          500:
            description: Internal Server Error